Create TranslationIntent
