+++ DailyWidget/TranslationEntry.swift
