+++ DailyWidget/TranslationView.swift
