+++ DailyWidget/TranslationWidget.swift
